nambah is_selected

+++ database/migrations/2023_10_17_193448_create_participant_criterias_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->integer('value')->nullable();
             
             $table->integer('weight');
-            $table->float('weight_normalization')->nullable();
+            $table->decimal('weight_normalization',10,9)->nullable();
             $table->longText('note')->nullable();
             $table->timestamps();
         });
